// Front-End/src/app/Models/User.ts
export class User {
  _id: string;
  accType: string;
  accStatus: string;
  accVerification: string;
  ProfileImage: string;
  email: string;
  FirstName: string;
  LastName: string;
  UserName: string;
  Password: string;
  Gender: string;
  Age: number;
  Location: string;
  PhoneNumber: number;
  CIN: number;
}